+++ src/editor/utils/element.ts
@@ -1240,7 +1240,7 @@ export function convertTextNodeToElement(
   return element
 }
 
-interface IGetElementListByHTMLOption {
+export interface IGetElementListByHTMLOption {
   innerWidth: number
 }